consolidates tests
